Re-write of the way data retrieval is handled
