+++ heinlein/dtypes/__init__.py
@@ -1,4 +1,4 @@
-from .handlers import get_handler
+from .handlers import get_handler, get_file_handlers
 from .dtypes import get_data_object
 
-__all__ = ["get_handler", "get_data_object"]
+__all__ = ["get_handler", "get_data_object", "get_file_handlers"]
