+++ heinlein/locations.py
@@ -11,6 +11,8 @@ MAIN_CONFIG_DIR = base / "config"
 BASE_DATASET_CONFIG_DIR = INSTALL_DIR / "dataset" / "configs"
 BASE_DATASET_CONFIG = BASE_DATASET_CONFIG_DIR / "surveys.json"
 
+BUILTIN_DTYPES = base / "dtypes"/ "dtypes.json"
+
 DATASET_CONFIG_DIR = data_dir / "datasets"
 DATASET_CONFIG_DIR.mkdir(parents=True, exist_ok=True)
 MAIN_DATASET_CONFIG = DATASET_CONFIG_DIR / "surveys.json"
