+++ heinlein/manager/__init__.py
@@ -1,4 +1,4 @@
-from .manager import FileManager, get_manager
+from .managers import get_manager
 import shutil
 
 from heinlein.locations import MAIN_DATASET_CONFIG, BASE_DATASET_CONFIG
@@ -7,4 +7,4 @@ if not MAIN_DATASET_CONFIG.exists():
     shutil.copy(BASE_DATASET_CONFIG, MAIN_DATASET_CONFIG)
 
 
-__all__ = ["FileManager", "get_manager"]
+__all__ = ["get_manager"]
